fix: regression: archive fails with nonexistent container
